+++ xclim/ensembles/_base.py
@@ -127,7 +127,7 @@ def create_ensemble(
 
 
 def ensemble_mean_std_max_min(
-    ens: xr.Dataset, weights: xr.DataArray | None = None
+    ens: xr.Dataset, min_members: int | None = None, weights: xr.DataArray | None = None
 ) -> xr.Dataset:
     """Calculate ensemble statistics between a results from an ensemble of climate simulations.
 
@@ -138,6 +138,9 @@ def ensemble_mean_std_max_min(
     ----------
     ens : xr.Dataset
         Ensemble dataset (see xclim.ensembles.create_ensemble).
+    min_members : int, optional
+        The minimum number of valid ensemble members for a statistic to be valid.
+        The default (None), is equivalent to setting min_members to the size of the realization dimension.
     weights : xr.DataArray, optional
         Weights to apply along the 'realization' dimension. This array cannot contain missing values.
 
@@ -170,9 +173,14 @@ def ensemble_mean_std_max_min(
         ds_out[f"{v}_max"] = ens[v].max(dim="realization")
         ds_out[f"{v}_min"] = ens[v].min(dim="realization")
 
+        if min_members is not None:
+            enough = ens[v].notnull().sum("realization") >= min_members
+
         # Re-add attributes
         for stat in ["mean", "stdev", "max", "min"]:
             vv = f"{v}_{stat}"
+            if min_members is not None:
+                ds_out[vv] = ds_out[vv].where(enough)
             ds_out[vv].attrs = ens[v].attrs
             if "description" in ds_out[vv].attrs.keys():
                 vv.split()
@@ -182,6 +190,7 @@ def ensemble_mean_std_max_min(
                     + vv.split("_")[-1]
                     + " of ensemble"
                 )
+
     ds_out.attrs["history"] = update_history(
         f"Computation of statistics on {ens.realization.size} ensemble members.", ds_out
     )
@@ -192,6 +201,7 @@ def ensemble_percentiles(
     ens: xr.Dataset | xr.DataArray,
     values: Sequence[int] | None = None,
     keep_chunk_size: bool | None = None,
+    min_members: int | None = None,
     weights: xr.DataArray | None = None,
     split: bool = True,
 ) -> xr.DataArray | xr.Dataset:
@@ -211,6 +221,9 @@ def ensemble_percentiles(
         so that the chunks keep the same size (approximately).
         If False, no shrinking is performed, resulting in much larger chunks.
         If not defined, the function decides which is best.
+    min_members : int, optional
+        The minimum number of valid ensemble members for a statistic to be valid.
+        The default (None), is equivalent to setting min_members to the size of the realization dimension.
     weights : xr.DataArray, optional
         Weights to apply along the 'realization' dimension. This array cannot contain missing values.
         When given, the function uses xarray's quantile method which is slower than xclim's NaN-optimized algorithm.
@@ -313,6 +326,8 @@ def ensemble_percentiles(
                 .rename({"quantile": "percentiles"})
             )
 
+    if min_members is not None:
+        out = out.where(ens.notnull().sum("realization") >= min_members)
     out = out.assign_coords(
         percentiles=xr.DataArray(list(values), dims=("percentiles",))
     )